+++ projects/demo/src/app/logger-form/logger-form.component.ts
@@ -1,8 +1,8 @@
-import {Component, OnInit, Output, EventEmitter} from '@angular/core';
-import {Validators, FormBuilder, FormGroup} from '@angular/forms';
-import {NGXLogger, NgxLoggerLevel} from 'ngx-logger';
+import { Component, OnInit, Output, EventEmitter } from '@angular/core';
+import { Validators, FormBuilder, FormGroup } from '@angular/forms';
+import { NGXLogger, NgxLoggerLevel } from '../../../../../src/public_api';
 
-import {LogEvent} from '../models/log-event.model';
+import { LogEvent } from '../models/log-event.model';
 
 export interface LoggerSelectionOption {
   value: NgxLoggerLevel;
@@ -28,12 +28,12 @@ export class LoggerFormComponent implements OnInit {
 
   /* Used in the mat-select on the form */
   logTypes: LoggerSelectionOption[] = [
-    {value: NgxLoggerLevel.TRACE, viewValue: 'Trace'},
-    {value: NgxLoggerLevel.DEBUG, viewValue: 'Debug'},
-    {value: NgxLoggerLevel.INFO, viewValue: 'Info'},
-    {value: NgxLoggerLevel.LOG, viewValue: 'Log'},
-    {value: NgxLoggerLevel.WARN, viewValue: 'Warn'},
-    {value: NgxLoggerLevel.ERROR, viewValue: 'Error'}
+    { value: NgxLoggerLevel.TRACE, viewValue: 'Trace' },
+    { value: NgxLoggerLevel.DEBUG, viewValue: 'Debug' },
+    { value: NgxLoggerLevel.INFO, viewValue: 'Info' },
+    { value: NgxLoggerLevel.LOG, viewValue: 'Log' },
+    { value: NgxLoggerLevel.WARN, viewValue: 'Warn' },
+    { value: NgxLoggerLevel.ERROR, viewValue: 'Error' }
   ];
 
   constructor(
@@ -53,9 +53,9 @@ export class LoggerFormComponent implements OnInit {
   }
 
   logComplex() {
-      const complexStructure = new FormGroup({ sub: new FormGroup({}) });
-      this.logger.error('Test complex', complexStructure);
-      this.logger.error(complexStructure);
+    const complexStructure = new FormGroup({ sub: new FormGroup({}) });
+    this.logger.error('Test complex', complexStructure);
+    this.logger.error(complexStructure);
   }
 
 }