+++ src/lib/logger.barrel.ts
@@ -1,15 +1,32 @@
-export * from './logger.module';
+export * from './config/iconfig';
+export * from './config/iconfig-engine';
+export * from './config/config-engine';
 
-export * from './logger.service';
-export * from './logger.config';
-export * from './custom-logger.service';
-export * from './logger-monitor';
-export * from './http.service';
-export * from './mapper.service';
+export * from './mapper/ilog-position';
+export * from './mapper/imapper-service';
+export * from './mapper/mapper-service';
 
-export * from './utils/logger.utils';
+export * from './metadata/imetadata-service';
+export * from './metadata/imetadata';
+export * from './metadata/metadata-service';
+
+export * from './monitor/ilogger-monitor';
+export * from './monitor/logger-monitor';
+
+export * from './rules/irules-service';
+export * from './rules/rules-service';
+
+export * from './server/iserver-service';
+export * from './server/server-service';
+
+export * from './writer/iwriter-service';
+export * from './writer/writer-service';
+export * from './writer/color-scheme';
+
 export * from './types/logger-level.enum';
 export * from './types/ngx-log.interface';
-export * from './types/logger-color-scheme';
 
-export * from './resources/color-schemes';
+export * from './logger.service';
+export * from './custom-logger.service';
+
+export * from './logger.module';