+++ src/lib/mapper/ilog-position.ts
@@ -1,0 +1,13 @@
+/**
+ * Position of caller
+ */
+export interface INGXLoggerLogPosition {
+    fileName?: string;
+    lineNumber?: number;
+    columnNumber?: number;
+
+    // toString() {
+    //     return this.fileName + ':' + this.lineNumber + ':' + this.columnNumber;
+    //   }
+
+}